Change double columns to float

+++ db/migrate/20161119031317_change_location_to_double.rb
@@ -1,6 +1,6 @@
 class ChangeLocationToDouble < ActiveRecord::Migration[5.0]
   def change
-    change_column :availabilities, :longitude, :double
-    change_column :availabilities, :latitude, :double
+    change_column :availabilities, :longitude, :float
+    change_column :availabilities, :latitude, :float
   end
 end
